Move Logstash labels into its own package
